backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended>=4.6.0
Flask-CORS==4.0.0
Flask-SocketIO>=5.3.0
Flask-Mail>=0.9.1
SQLAlchemy>=2.0.23
python-dotenv==1.0.0
bcrypt==4.1.2
pyotp==2.9.0
qrcode[pil]==7.4.2
Pillow>=9.0.0
itsdangerous>=2.1.2
alembic>=1.13.0
psycopg2-binary>=2.9.0
requests==2.31.0
pytest>=7.4.2
pytest-cov>=4.1.0
hypothesis>=6.82.0
python-dateutil>=2.8.2
gunicorn>=21.2.0
openpyxl>=3.1.2
pyzk>=0.9
APScheduler>=3.10.0
pusher>=3.3.2
